修复d03ssh02 MAC拼写错误．将d05compile2改为测试机器d05ssh09,现在有9台d05测试机

+++ test-scripts/jenkins_boot_start.sh
@@ -28,6 +28,8 @@ function init_input_params() {
 
     GIT_DESCRIBE=${GIT_DESCRIBE:-""}
 
+    DEBUG=${DEBUG:-""}
+
     JENKINS_JOB_INFO=$(expr "${BUILD_URL}" : '^http.*/job/\(.*\)/$' | sed "s#/#-#g")
 
     JENKINS_JOB_START_TIME=${JENKINS_JOB_START_TIME:-$(current_time)}
@@ -78,6 +80,7 @@ BOOT_PLAN=${BOOT_PLAN}
 TEST_REPO=${TEST_REPO}
 TEST_SCOPE="${TEST_SCOPE}"
 TEST_LEVEL=${TEST_LEVEL}
+DEBUG=${DEBUG}
 JENKINS_JOB_START_TIME="${JENKINS_JOB_START_TIME}"
 ARCH_MAP="${ARCH_MAP}"
 EOF
@@ -379,12 +382,14 @@ function collect_result() {
     #zip -r ${{GIT_DESCRIBE}}_results.zip ${GIT_DESCRIBE}/*
     cp -f ${timefile} ${GIT_DESCRIBE} || true
 
-    if [ -d $DES_DIR/${GIT_DESCRIBE}/results ];then
-        sudo rm -fr $DES_DIR/${GIT_DESCRIBE}/results
-        sudo rm -fr $DES_DIR/${GIT_DESCRIBE}/${timefile}
+    # if DEBUG is set, don't update the result
+    if [ ! "${DEBUG}" = "true" ];then
+        if [ -d $DES_DIR/${GIT_DESCRIBE}/results ];then
+            sudo rm -fr $DES_DIR/${GIT_DESCRIBE}/results
+            sudo rm -fr $DES_DIR/${GIT_DESCRIBE}/${timefile}
+        fi
+        sudo cp -rf ${GIT_DESCRIBE}/* $DES_DIR
     fi
-
-    sudo cp -rf ${GIT_DESCRIBE}/* $DES_DIR
 
     popd    # restore current work directory
 
@@ -453,13 +458,9 @@ function generate_success_mail(){
 
     # prepare parameters
     cd ${WORKSPACE}
-    if [ "${DEBUG}" = "true" ];then
-        echo "${FAILED_MAIL_LIST}" > ${WORKSPACE}/MAIL_LIST.txt
-        echo "${FAILED_MAIL_CC_LIST}" > ${WORKSPACE}/MAIL_CC_LIST.txt
-    else
-        echo "${SUCCESS_MAIL_LIST}" > ${WORKSPACE}/MAIL_LIST.txt
-        echo "${SUCCESS_MAIL_CC_LIST}" > ${WORKSPACE}/MAIL_CC_LIST.txt
-    fi
+    echo "${SUCCESS_MAIL_LIST}" > ${WORKSPACE}/MAIL_LIST.txt
+    echo "${SUCCESS_MAIL_CC_LIST}" > ${WORKSPACE}/MAIL_CC_LIST.txt
+
     TODAY=$(date +"%Y/%m/%d")
     MONTH=$(date +"%Y%m")
